+++ src/components/RangeField/Thumb/index.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { ReactSliderProps } from 'react-slider';
+
+import { StyledThumb } from './Thumb.styled';
+
+export const Thumb: ReactSliderProps['renderThumb'] = (props) => (
+  <StyledThumb size={6} thumbColorToken="white100" {...(props as object)} />
+);
